requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
pdfminer.six>=20221105
PyPDF2>=3.0.0
plotly>=5.17.0
python-dotenv>=1.0.0